onprem-app/service-a/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
pydantic
httpx
python-multipart